+++ src/ui/settings-dialog.h
@@ -58,6 +58,7 @@ private:
 	QComboBox *scene_ = nullptr, *game_ = nullptr, *gameDetect_ = nullptr;
 	QTableWidget *friends_;
 	QPushButton *speedBtn_ = nullptr;
+	QLabel *ndiStatus_ = nullptr;
 	QComboBox *lookPos_ = nullptr, *ndiQuality_ = nullptr;
 	QListWidget *onTop_ = nullptr;
 	QListWidget *mute_ = nullptr;